+++ docs/source/oauth2client.contrib.xsrfutil.rst
@@ -1,0 +1,7 @@
+oauth2client.contrib.xsrfutil module
+====================================
+
+.. automodule:: oauth2client.contrib.xsrfutil
+    :members:
+    :undoc-members:
+    :show-inheritance:
