--- docs/source/oauth2client.xsrfutil.rst
@@ -1,7 +1,0 @@
-oauth2client.xsrfutil module
-============================
-
-.. automodule:: oauth2client.xsrfutil
-    :members:
-    :undoc-members:
-    :show-inheritance:
