+++ oauth2client/contrib/appengine.py
@@ -38,13 +38,13 @@ from oauth2client import GOOGLE_REVOKE_URI
 from oauth2client import GOOGLE_TOKEN_URI
 from oauth2client import clientsecrets
 from oauth2client import util
-from oauth2client import xsrfutil
 from oauth2client.client import AccessTokenRefreshError
 from oauth2client.client import AssertionCredentials
 from oauth2client.client import Credentials
 from oauth2client.client import Flow
 from oauth2client.client import OAuth2WebServerFlow
 from oauth2client.client import Storage
+from oauth2client.contrib import xsrfutil
 
 # TODO(dhermes): Resolve import issue.
 # This is a temporary fix for a Google internal issue.
