+++ tests/contrib/test_xsrfutil.py
@@ -11,9 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tests for oauth2client.xsrfutil.
+"""Tests for oauth2client.contrib.xsrfutil.
 
-Unit tests for oauth2client.xsrfutil.
+Unit tests for oauth2client.contrib.xsrfutil.
 """
 
 import base64
@@ -22,7 +22,7 @@ import unittest
 import mock
 
 from oauth2client._helpers import _to_bytes
-from oauth2client import xsrfutil
+from oauth2client.contrib import xsrfutil
 
 # Jan 17 2008, 5:40PM
 TEST_KEY = b'test key'
@@ -52,7 +52,7 @@ class Test_generate_token(unittest.TestCase):
         curr_time = 1440449755.74
         digester = mock.MagicMock()
         digester.digest = mock.MagicMock(name='digest', return_value=digest)
-        with mock.patch('oauth2client.xsrfutil.hmac') as hmac:
+        with mock.patch('oauth2client.contrib.xsrfutil.hmac') as hmac:
             hmac.new = mock.MagicMock(name='new', return_value=digester)
             token = xsrfutil.generate_token(TEST_KEY,
                                             TEST_USER_ID_1,
@@ -81,10 +81,10 @@ class Test_generate_token(unittest.TestCase):
         curr_time = 1440449755.74
         digester = mock.MagicMock()
         digester.digest = mock.MagicMock(name='digest', return_value=digest)
-        with mock.patch('oauth2client.xsrfutil.hmac') as hmac:
+        with mock.patch('oauth2client.contrib.xsrfutil.hmac') as hmac:
             hmac.new = mock.MagicMock(name='new', return_value=digester)
 
-            with mock.patch('oauth2client.xsrfutil.time') as time:
+            with mock.patch('oauth2client.contrib.xsrfutil.time') as time:
                 time.time = mock.MagicMock(name='time', return_value=curr_time)
                 # when= is omitted
                 token = xsrfutil.generate_token(TEST_KEY,
@@ -140,7 +140,7 @@ class Test_validate_token(unittest.TestCase):
 
         key = user_id = None
         token = base64.b64encode(_to_bytes(str(token_time)))
-        with mock.patch('oauth2client.xsrfutil.time') as time:
+        with mock.patch('oauth2client.contrib.xsrfutil.time') as time:
             time.time = mock.MagicMock(name='time', return_value=curr_time)
             self.assertFalse(xsrfutil.validate_token(key, token, user_id))
             time.time.assert_called_once_with()
@@ -167,7 +167,7 @@ class Test_validate_token(unittest.TestCase):
         # Make sure the token length comparison will fail.
         self.assertNotEqual(len(token), len(generated_token))
 
-        with mock.patch('oauth2client.xsrfutil.generate_token',
+        with mock.patch('oauth2client.contrib.xsrfutil.generate_token',
                         return_value=generated_token) as gen_tok:
             self.assertFalse(xsrfutil.validate_token(key, token, user_id,
                                                      current_time=curr_time,
@@ -191,7 +191,7 @@ class Test_validate_token(unittest.TestCase):
         self.assertEqual(len(token), len(generated_token))
         self.assertNotEqual(token, generated_token)
 
-        with mock.patch('oauth2client.xsrfutil.generate_token',
+        with mock.patch('oauth2client.contrib.xsrfutil.generate_token',
                         return_value=generated_token) as gen_tok:
             self.assertFalse(xsrfutil.validate_token(key, token, user_id,
                                                      current_time=curr_time,
@@ -208,7 +208,7 @@ class Test_validate_token(unittest.TestCase):
         user_id = object()
         action_id = object()
         token = base64.b64encode(_to_bytes(str(token_time)))
-        with mock.patch('oauth2client.xsrfutil.generate_token',
+        with mock.patch('oauth2client.contrib.xsrfutil.generate_token',
                         return_value=token) as gen_tok:
             self.assertTrue(xsrfutil.validate_token(key, token, user_id,
                                                     current_time=curr_time,
